# docs/api.yaml
openapi: 3.0.0
info:
  title: PrepAIr API
  description: AI-Powered Career Preparation Platform API
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /api/cv/ingest:
    post:
      summary: Ingest CV text
      tags: [cv]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, cv_text]
              properties:
                user_id:
                  type: string
                cv_text:
                  type: string
      responses:
        '200':
          description: CV version created
          content:
            application/json:
              schema:
                type: object
                properties:
                  cv_version_id:
                    type: string
  /api/cv/analyze:
    post:
      summary: Analyze CV against job spec
      tags: [cv]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, cv_version_id, job_spec_id]
              properties:
                user_id:
                  type: string
                cv_version_id:
                  type: string
                job_spec_id:
                  type: string
      responses:
        '200':
          description: Analysis results
          content:
            application/json:
              schema:
                type: object
                properties:
                  match_score:
                    type: number
                  strengths:
                    type: array
                    items:
                      type: string
                  gaps:
                    type: array
                    items:
                      type: string
                  suggestions:
                    type: array
                    items:
                      type: string
                  role_focus:
                    type: object
  /api/cv/save:
    post:
      summary: Save improved CV version
      tags: [cv]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, updated_cv_text]
              properties:
                user_id:
                  type: string
                parent_cv_version_id:
                  type: string
                updated_cv_text:
                  type: string
      responses:
        '200':
          description: New CV version ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  new_cv_version_id:
                    type: string
...